Reorganize utils folder.
